+++ src/components/handshakes/betting/Feed/BetHandshakeHandler.js
@@ -1,186 +1,16 @@
-import { MasterWallet } from '@/services/Wallets/MasterWallet';
 import { BettingHandshake } from '@/services/neuron';
-import { API_URL, APP } from '@/constants';
+import { API_URL } from '@/constants';
 import { showAlert } from '@/reducers/app/action';
-import { PredictionHandshake } from '@/services/neuron';
-import { getMessageWithCode, getChainIdDefaultWallet, isInitBet, } from '@/components/handshakes/betting/utils.js';
-
+import { getMessageWithCode, getChainIdDefaultWallet, isInitBet } from '@/components/handshakes/betting/utils.js';
+import { MESSAGE, BETTING_STATUS_LABEL } from '@/components/handshakes/betting/message.js';
 import { isExpiredDate } from '@/components/handshakes/betting/validation.js';
-
-import Web3 from 'web3';
+import { BET_BLOCKCHAIN_STATUS, ROLE, SIDE, BETTING_RESULT, CONTRACT_METHOD } from '@/components/handshakes/betting/constants.js';
 import GA from '@/services/googleAnalytics';
 
-import local from '@/services/localStore';
-import { rollback, saveTransaction, collect } from '@/reducers/handshake/action';
+import { rollback, saveTransaction } from '@/reducers/handshake/action';
 
 import store from '@/stores';
-import moment from 'moment';
-import { off } from 'rsvp';
 
-export const MESSAGE_SERVER = {
-  /* ERROR */
-  1000: 'Please double check your input data.',
-  1001: 'No equivalent bets found. Please create a new bet.',
-  1002: 'Please provide valid wallet address!',
-  1003: 'Missing offchain data!',
-  1004: 'Odds should be large than 1',
-  1005: 'The maximum free bet is 100!',
-  1006: 'You cannot withdraw this handshake!',
-  1007: 'Cannot rollback this handshake!',
-
-  /* OUTCOME */
-  1008: 'Please check your outcome id',
-  1009: 'Please check your outcome result',
-  1010: 'This outcome has had result already!',
-
-  /* MATCH */
-  1011: 'Match not found. Please try again.',
-  1012: 'Match result invalid. Please try again.',
-  1013: 'Match result is empty. Please try again.',
-  1014: 'The report time is exceed!',
-
-  /* USER */
-  1015: 'Please enter a valid email address.',
-  1016: 'Please make sure your email and password are correct.',
-  1017: 'Sorry, we were unable to register you. Please contact human@autonomous.ai for support.',
-  1018: 'Invalid user',
-  1019: 'Please purchase to sign more.',
-  1020: 'Invalid user',
-  1021: 'Please login with google+ or facebook account.',
-  1022: 'You have received free bet already!',
-
-  /* HANSHAKE */
-  1023: 'You\'re out of gas! Please wait while we add ETH to your account.',
-  1024: 'You can\'t Handshake with yourself!',
-  1025: 'This Handshake seems to be empty.',
-  1026: 'You are not authorized to make this Handshake.',
-  1027: 'Contract file not found!',
-  1028: 'Handshake not found. Please try again.',
-  1029: 'Please enter a payment amount.',
-  1030: 'Amount should be larger > 0.',
-  1031: 'Amount key is invalid.',
-  1032: 'Public key is invalid.',
-  1033: 'Please enter a valid wallet address which exists in our system.',
-  1034: 'You\'re out of gas! Please wait while we add ETH to your account.',
-  1035: 'Your note is too long. It should be less than 1000 characters.',
-  1036: 'Please choose type of handshake.',
-  1037: 'This is not betting template.',
-  1038: 'There is shakers. Therefore you cannot refund!',
-  1039: 'Your result does not match with outcome!',
-  1040: 'Withdraw only works after dispute time.',
-
-  /* SHAKER */
-  1041: 'Shaker not found. Please try again.',
-  1042: 'You have rollbacked already!',
-
-
-  /* WALLET */
-  1043: 'Busy day for Handshakes - we\'re out of freebies! Please try again tomorrow.',
-  1044: 'You can only request free Handshakes once.',
-  1045: 'Your account can\'t get free ETH.',
-};
-
-
-export const MESSAGE = {
-  BET_PROGRESSING: 'Your bet is creating. Please wait',
-  CREATE_BET_NOT_MATCH: 'Finding a ninja to bet against you.',
-  CREATE_BET_MATCHED: 'Bet matched! Waiting for outcome.',
-  NOT_ENOUGH_BALANCE: 'Too rich for your blood. Please top up your wallet.',
-  NOT_ENOUGH_GAS: `Not enough gas. Your balance should larger than 0.007eth gas + value. Please top up your wallet.`,
-  CHOOSE_MATCH: 'Please choose event and outcome',
-  ODD_LARGE_THAN: 'Please enter odds greater than 1 and smaller 12',
-  AMOUNT_VALID: 'Please place a bet larger than 0.',
-  MATCH_OVER: 'Time travel is hard. Please bet on a future or ongoing match.',
-  RIGHT_NETWORK: 'You must set your wallet on Mainnet',
-  ROLLBACK: `Something did not go according to plan. Please try again.`,
-  WITHDRAW_SUCCESS: 'Success! Your winnings have been withdrawn to your wallet.',
-  DIFFERENCE_ADDRESS: `Current address isn't same as which you used to create bet`,
-};
-
-
-export const BET_BLOCKCHAIN_STATUS = {
-  STATUS_MAKER_SHOULD_UNINIT: -12,
-  STATUS_INIT_FAILED: -10,
-  STATUS_COLLECT_FAILED: -9,
-  STATUS_COLLECT_PENDING: -8,
-  STATUS_DISPUTE_FAILED: -7,
-  STATUS_REFUND_FAILED: -6,
-  STATUS_MAKER_UNINIT_FAILED: -5,
-  STATUS_MAKER_UNINIT_PENDING: -4,
-
-  STATUS_SHAKER_ROLLBACK: -3,
-  STATUS_MAKER_INIT_ROLLBACK: -2,
-
-
-  STATUS_PENDING: -1,
-  STATUS_INITED: 0,
-  STATUS_MAKER_UNINITED: 1,
-  STATUS_SHAKER_SHAKED: 2,
-  STATUS_REFUND: 3,
-  STATUS_DISPUTE: 4,
-  STATUS_RESOLVE: 5,
-  STATUS_DONE: 6,
-};
-
-export const ROLE = {
-  INITER: 1,
-  SHAKER: 2,
-};
-
-export const SIDE = {
-  SUPPORT: 1,
-  AGAINST: 2,
-};
-export const BET_TYPE = {
-  INIT: 'init',
-  SHAKE: 'shake',
-};
-export const BETTING_STATUS = {
-  INITED: -1,
-  DRAW: 0,
-  SUPPORT_WIN: 1,
-  AGAINST_WIN: 2,
-};
-
-export const BETTING_STATUS_LABEL =
-    {
-      INITING: 'Placing a bet..',
-      CANCEL: 'Cancel this bet',
-      RETRY: 'Retry',
-      ROLLBACK_INIT: 'There is something wrong with blockchain. The bet is cancelled',
-      ROLLBACK_SHAKE: 'There is something wrong with blockchain. The bet is cancelled',
-      COLLECT_FAILED: 'There is something wrong with withdraw. Please cancel to get back money',
-      SHOULD_CANCEL: 'There is no bet matched. Please cancel to get back money',
-      MATCH_POSTPONSE: 'Event is postponsed',
-      ACTION_FAILED: `There is something wrong with blockchain. Your action is cancelled`,
-      INIT_FAILED: `There is something wrong with blockchain. Your bet is cancelled`,
-      SOLVE: 'Please retry to solve problem',
-      LOSE: 'Better luck next time.',
-      WIN: `You're a winner!`,
-      WIN_WAIT: `<br/> It's not time to withdraw. Please wait..`,
-      DONE: ` The bet is done.`,
-      WITHDRAW: 'Withdraw winnings',
-      CANCELLING: 'Your bet is being cancelled.',
-      PROGRESSING: 'Your bet is progressing a transaction. Please wait..',
-      BET_WAIT_MATCHING: 'Bet placed. Matching..',
-      BET_MACHED_WAIT_RESULT: 'Bet matched. Waiting for result..',
-      REFUND: 'Refund your bet',
-      CANCELLED: 'Your bet was cancelled.',
-      REFUNDING: 'Your coin is being refunded to you.',
-      REFUNDED: 'Your coin has been refunded.',
-      ROLLBACK: `Something did not go according to plan. We're fixing it`,
-      COLLECT_PENDING: 'The blockchain is processing your withdrawal. Please wait.',
-      COLLECT_DONE: 'Your winnings have been withdrawn to your wallet!',
-    };
-
-export const CONTRACT_METHOD = {
-  INIT: 'init',
-  SHAKE: 'shake',
-  CANCEL: 'uninit',
-  REFUND: 'refund',
-  COLLECT: 'collect',
-  CREATE_MARKET: 'createMarket',
-};
 
 const myManager = null;
 
@@ -212,10 +42,10 @@ export class BetHandshakeHandler {
       label = BETTING_STATUS_LABEL.CANCEL;
       strStatus = BETTING_STATUS_LABEL.COLLECT_FAILED;
       isAction = true;
-    } else if (isMatch && resultStatus === BETTING_STATUS.DRAW) {
-      //label = BETTING_STATUS_LABEL.CANCEL;
+    } else if (isMatch && resultStatus === BETTING_RESULT.DRAW) {
+      label = BETTING_STATUS_LABEL.REFUND;
       strStatus = BETTING_STATUS_LABEL.MATCH_POSTPONSE;
-      isAction = false;
+      isAction = true;
     } else if (blockchainStatus === BET_BLOCKCHAIN_STATUS.STATUS_MAKER_SHOULD_UNINIT) {
       label = BETTING_STATUS_LABEL.CANCEL;
       strStatus = BETTING_STATUS_LABEL.SHOULD_CANCEL;
@@ -246,20 +76,20 @@ export class BetHandshakeHandler {
     } else if (blockchainStatus === BET_BLOCKCHAIN_STATUS.STATUS_REFUND) {
       strStatus = BETTING_STATUS_LABEL.REFUNDED;
       isAction = false;
-    } else if ((blockchainStatus === BET_BLOCKCHAIN_STATUS.STATUS_DONE && resultStatus === BETTING_STATUS.SUPPORT_WIN && side === SIDE.SUPPORT)
-                || (blockchainStatus === BET_BLOCKCHAIN_STATUS.STATUS_DONE && resultStatus === BETTING_STATUS.AGAINST_WIN && side === SIDE.AGAINST)) {
+    } else if ((blockchainStatus === BET_BLOCKCHAIN_STATUS.STATUS_DONE && resultStatus === BETTING_RESULT.SUPPORT_WIN && side === SIDE.SUPPORT)
+                || (blockchainStatus === BET_BLOCKCHAIN_STATUS.STATUS_DONE && resultStatus === BETTING_RESULT.AGAINST_WIN && side === SIDE.AGAINST)) {
       strStatus = BETTING_STATUS_LABEL.COLLECT_DONE;
       isAction = false;
     } else if (!isMatch && role === ROLE.INITER && blockchainStatus !== BET_BLOCKCHAIN_STATUS.STATUS_SHAKER_SHAKED) {
       label = BETTING_STATUS_LABEL.CANCEL;
       strStatus = BETTING_STATUS_LABEL.BET_WAIT_MATCHING;
       isAction = true;
-    } else if (isMatch && resultStatus === BETTING_STATUS.INITED && isExpiredDate(reportTime)) {
+    } else if (isMatch && resultStatus === BETTING_RESULT.INITED && isExpiredDate(reportTime)) {
       label = BETTING_STATUS_LABEL.REFUND;
       strStatus = BETTING_STATUS_LABEL.REFUNDING;
       isAction = true;
-    } else if ((isMatch && resultStatus === BETTING_STATUS.SUPPORT_WIN && side === SIDE.SUPPORT)
-                || (isMatch && resultStatus === BETTING_STATUS.AGAINST_WIN && side === SIDE.AGAINST)) {
+    } else if ((isMatch && resultStatus === BETTING_RESULT.SUPPORT_WIN && side === SIDE.SUPPORT)
+                || (isMatch && resultStatus === BETTING_RESULT.AGAINST_WIN && side === SIDE.AGAINST)) {
       if (isExpiredDate(disputeTime)) { // Over dispute time, user can withdraw
         label = BETTING_STATUS_LABEL.WITHDRAW;
         strStatus = BETTING_STATUS_LABEL.WIN;
@@ -268,8 +98,8 @@ export class BetHandshakeHandler {
         strStatus = BETTING_STATUS_LABEL.WIN + BETTING_STATUS_LABEL.WIN_WAIT;
         isAction = true;
       }
-    } else if ((isMatch && resultStatus === BETTING_STATUS.SUPPORT_WIN && side === SIDE.AGAINST)
-                || (isMatch && resultStatus === BETTING_STATUS.AGAINST_WIN && side === SIDE.SUPPORT)) {
+    } else if ((isMatch && resultStatus === BETTING_RESULT.SUPPORT_WIN && side === SIDE.AGAINST)
+                || (isMatch && resultStatus === BETTING_RESULT.AGAINST_WIN && side === SIDE.SUPPORT)) {
       strStatus = BETTING_STATUS_LABEL.LOSE;
       isAction = false;
     } else if (isMatch || blockchainStatus === BET_BLOCKCHAIN_STATUS.STATUS_SHAKER_SHAKED) {
@@ -297,7 +127,7 @@ export class BetHandshakeHandler {
       dataBlockchain = await bettinghandshake.initBet(hid, side, stake, odds, offchain);
       // TO DO: SAVE TRANSACTION
       const {
-        logs, hash, error, transactionHash, payload,
+        hash, error, payload,
       } = dataBlockchain;
       logJson = payload;
       realBlockHash = hash;
@@ -336,7 +166,7 @@ export class BetHandshakeHandler {
     console.log('shakeContract', item);
 
     const {
-      amount, id, odds, side, maker_address, maker_odds, offchain, hid,
+      amount, odds, side, maker_address, maker_odds, offchain, hid,
     } = item;
     // hid = 10000;
     const stake = Math.floor(amount * 10 ** 18) / 10 ** 18;
@@ -360,7 +190,7 @@ export class BetHandshakeHandler {
         offchain,
       );
       const {
-        logs, hash, error, transactionHash, payload,
+        hash, error, payload,
       } = result;
 
       logJson = payload;
@@ -396,8 +226,6 @@ export class BetHandshakeHandler {
 
   handleContract(element, i) {
     setTimeout(() => {
-      console.log('Time out:');
-      const { offchain, odds } = element;
       const isInit = isInitBet(element);
       console.log('Is Init Bet:', isInit);
       if (isInit) {
@@ -406,24 +234,8 @@ export class BetHandshakeHandler {
         this.shakeContract(element);
       }
     }, 3000 * i);
-
-
-    /*
-   const { offchain, odds } = element;
-   const isInit = isInitBet(element);
-   console.log('Is Init Bet:', isInit);
-   if (isInit) {
-     this.addContract(element);
-   } else {
-
-   this.shakeContract(element);
-
-   }
-   */
   }
   controlShake = async (list) => {
-    const result = null;
-
     for (let i = 0; i < list.length; i++) {
       const element = list[i];
       console.log('Element:', element);
@@ -445,7 +257,7 @@ export class BetHandshakeHandler {
     try {
       result = await bettinghandshake.cancelBet(hid, side, stake, odds, offchain);
       const {
-        logs, hash, error, transactionHash, payload,
+         hash, error, payload,
       } = result;
 
       logJson = payload;
@@ -495,7 +307,7 @@ export class BetHandshakeHandler {
     try {
       result = await bettinghandshake.withdraw(hid, offchain);
       const {
-        logs, hash, error, transactionHash, payload,
+        hash, error, payload,
       } = result;
       logJson = payload;
       realBlockHash = hash;
