+++ src/components/handshakes/betting/Feed/Feed.jsx
@@ -3,22 +3,20 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 // services, constants
-import { BetHandshakeHandler, SIDE, BETTING_STATUS_LABEL, ROLE, MESSAGE, BET_BLOCKCHAIN_STATUS, BETTING_STATUS } from './BetHandshakeHandler.js';
-import momment from 'moment';
-import { MasterWallet } from '@/services/Wallets/MasterWallet';
+import { BetHandshakeHandler } from '@/components/handshakes/betting/Feed/BetHandshakeHandler';
+import { BET_BLOCKCHAIN_STATUS, ROLE, SIDE, BETTING_RESULT } from '@/components/handshakes/betting/constants.js';
+
 
 import local from '@/services/localStore';
-import { FIREBASE_PATH, HANDSHAKE_ID, API_URL, APP } from '@/constants';
+import { API_URL, APP } from '@/constants';
 import { uninitItem, collect, refund, collectFree, rollback, uninitItemFree } from '@/reducers/handshake/action';
 import { loadMyHandshakeList, updateBettingChange } from '@/reducers/me/action';
+import { MESSAGE, BETTING_STATUS_LABEL } from '@/components/handshakes/betting/message.js';
 
 
 // components
-import Image from '@/components/core/presentation/Image';
 import Button from '@/components/core/controls/Button';
-import ModalDialog from '@/components/core/controls/ModalDialog';
 import Feed from '@/components/core/presentation/Feed';
-import BettingShake from './Shake';
 import { showAlert } from '@/reducers/app/action';
 import {
   getMessageWithCode, getId,
@@ -27,34 +25,12 @@ import {
 
 import { isSameAddress, isRightNetwork } from '@/components/handshakes/betting/validation.js';
 
-
-
 // css, icons
 import './Feed.scss';
-import chipIcon from '@/assets/images/icon/betting/chip.svg';
-import conferenceCallIcon from '@/assets/images/icon/betting/conference_call.svg';
-import ethereumIcon from '@/assets/images/icon/betting/ethereum.svg';
-import Shake from '@/components/core/controls/Button';
-
-import GroupBook from './GroupBook';
 
 const betHandshakeHandler = BetHandshakeHandler.getShareManager();
 const ROUND = 1000000;
 const ROUND_ODD = 10;
-const BACKGROUND_COLORS = [
-  'linear-gradient(-135deg, #FFA7E7 0%, #EA6362 100%)',
-  'linear-gradient(-135deg, #17EAD9 0%, #6078EA 100%)',
-  'linear-gradient(-135deg, #23BCBA 0%, #45E994 100%)',
-  'linear-gradient(-135deg, #FFDEA7 0%, #EA6362 100%)',
-  'linear-gradient(-133deg, #9B3CB7 0%, #FF396F 100%)',
-  'linear-gradient(-133deg, #004B91 0%, #78CC37 100%)',
-  'linear-gradient(-135deg, #38B8F2 0%, #843CF6 100%)',
-  'linear-gradient(-135deg, #E35C67 0%, #381CE2 100%)',
-  'linear-gradient(-135deg, #EFBFD5 0%, #9D61FD 100%)',
-  'linear-gradient(-135deg, #45E0A7 0%, #D5E969 100%)',
-  'linear-gradient(45deg, #FBC79A 0%, #D73E68 100%)',
-  'linear-gradient(45deg, #F6AB3E 0%, #8137F7 100%)',
-];
 
 class FeedBetting extends React.Component {
   backgroundCss(background) {
@@ -135,8 +111,8 @@ class FeedBetting extends React.Component {
 
     if (isMakerShakerSameUser) {
       // Is Maker win ?
-      if ((isMatch && result === BETTING_STATUS.SUPPORT_WIN && side === SIDE.SUPPORT)
-          || (isMatch && result === BETTING_STATUS.AGAINST_WIN && side === SIDE.AGAINST)) {
+      if ((isMatch && result === BETTING_RESULT.SUPPORT_WIN && side === SIDE.SUPPORT)
+          || (isMatch && result === BETTING_RESULT.AGAINST_WIN && side === SIDE.AGAINST)) {
         isUserShake = false;
       }
     }
